Sound and mic adjustment added to waybar

+++ waybar/launch.sh
@@ -1,7 +1,30 @@
-# quiting all waybar instances
-killall waybar
+#!/bin/bash
 
-# laoding waybar configuration
+# start waybar if not started
+if ! pgrep -x "waybar" > /dev/null; then
+	waybar &
+fi
 
-waybar -c ~/.config/waybar & waybar -s ~/.config/waybar/style.css
+# current checksums
+current_checksum_config=$(md5sum ~/.config/waybar/config)
+current_checksum_style=$(md5sum ~/.config/waybar/style.css)
 
+# loop forever
+while true; do
+	# new checksums
+	new_checksum_config=$(md5sum ~/.config/waybar/config)
+	new_checksum_style=$(md5sum ~/.config/waybar/style.css)
+
+	# if checksums are different
+	if [ "$current_checksum_config" != "$new_checksum_config" ] || [ "$current_checksum_style" != "$new_checksum_style" ]; then
+		# kill waybar
+		killall waybar
+
+		# start waybar
+		waybar &
+
+		# update checksums
+		current_checksum_config=$new_checksum_config
+		current_checksum_style=$new_checksum_style
+	fi
+done
